use XPending()
the reason for not using XPending() to begin with stems from a
misunderstanding of how it worked. i was under the impression that
XPending always flushes the buffer and tries to read more events.

however, that's not really the case after reading the manpage a bit more
carefully. XPending() flushes the buffer and attempts to read only when
there's no events in the queue already.

in other words, XPending does what i was manually doing with
XEventsQueued(). so go ahead and use XPending since it's shorter.