ungrab mouse and keyboard early and always
ungrabbing it early can make for easier debugging.

and it makes sense to ungrab it always instead of only in debug
builds since keeping the mouse/keyboard locked up if IO was
taking long (e.g writing to a pipe without any reader on the
opposite end) makes no sense.